api_base_app/requirements.txt
absl-py==1.0.0
asgiref==3.5.0
astunparse==1.6.3
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
ci-info==0.2.0
click==8.0.4
colorama==0.4.4
configobj==5.0.6
configparser==5.2.0
Django==4.0.3
djangorestframework==3.13.1
filelock==3.6.0
flatbuffers==2.0
future==0.18.2
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.2
isodate==0.6.1
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
lxml==4.8.0
Markdown==3.3.6
networkx==2.7
numpy==1.22.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pathlib==1.0.1
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
PyMuPDF==1.19.5
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
pyxnat==1.4
rdflib==6.1.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
traits==6.3.2
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
